// src/api/models/service-request-creation-model.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Argento Oficina Virtual
 * Api para Portal de Clientes Argento
 *
 * The version of the OpenAPI document: 0.10.2
 * Contact: info@mts.com.do
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { ContactInfoModel } from './contact-info-model';
// May contain unused imports in some cases
// @ts-ignore
import { FileUploadModel } from './file-upload-model';
// May contain unused imports in some cases
// @ts-ignore
import { ServiceRequestClientIdentificationModel } from './service-request-client-identification-model';
// May contain unused imports in some cases
// @ts-ignore
import { ServiceRequestClientModel } from './service-request-client-model';
// May contain unused imports in some cases
// @ts-ignore
import { ServiceRequestCreationModelFormData } from './service-request-creation-model-form-data';

/**
 * 
 * @export
 * @interface ServiceRequestCreationModel
 */
export interface ServiceRequestCreationModel {
    /**
     * 
     * @type {ServiceRequestClientModel}
     * @memberof ServiceRequestCreationModel
     */
    'client'?: ServiceRequestClientModel;
    /**
     * 
     * @type {ServiceRequestClientIdentificationModel}
     * @memberof ServiceRequestCreationModel
     */
    'clientIdentification'?: ServiceRequestClientIdentificationModel;
    /**
     * 
     * @type {Array<ContactInfoModel>}
     * @memberof ServiceRequestCreationModel
     */
    'clientContactInfo'?: Array<ContactInfoModel> | null;
    /**
     * 
     * @type {string}
     * @memberof ServiceRequestCreationModel
     */
    'formType'?: string | null;
    /**
     * 
     * @type {ServiceRequestCreationModelFormData}
     * @memberof ServiceRequestCreationModel
     */
    'formData'?: ServiceRequestCreationModelFormData;
    /**
     * 
     * @type {boolean}
     * @memberof ServiceRequestCreationModel
     */
    'ignoreBTypeVisibilityValidation'?: boolean;
    /**
     * 
     * @type {number}
     * @memberof ServiceRequestCreationModel
     */
    'typeId'?: number;
    /**
     * 
     * @type {number}
     * @memberof ServiceRequestCreationModel
     */
    'businessOwnerId'?: number | null;
    /**
     * 
     * @type {string}
     * @memberof ServiceRequestCreationModel
     */
    'responsibleUserId'?: string | null;
    /**
     * 
     * @type {Array<FileUploadModel>}
     * @memberof ServiceRequestCreationModel
     */
    'files'?: Array<FileUploadModel> | null;
    /**
     * 
     * @type {boolean}
     * @memberof ServiceRequestCreationModel
     */
    'sendEmail'?: boolean;
}