// src/api/models/service-request-edition-model.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Argento Oficina Virtual
 * Api para Portal de Clientes Argento
 *
 * The version of the OpenAPI document: 0.10.2
 * Contact: info@mts.com.do
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { ContactInfoModel } from './contact-info-model';
// May contain unused imports in some cases
// @ts-ignore
import { ServiceRequestClientIdentificationModel } from './service-request-client-identification-model';
// May contain unused imports in some cases
// @ts-ignore
import { ServiceRequestClientModel } from './service-request-client-model';
// May contain unused imports in some cases
// @ts-ignore
import { ServiceRequestCreationModelFormData } from './service-request-creation-model-form-data';

/**
 * 
 * @export
 * @interface ServiceRequestEditionModel
 */
export interface ServiceRequestEditionModel {
    /**
     * 
     * @type {ServiceRequestClientModel}
     * @memberof ServiceRequestEditionModel
     */
    'client'?: ServiceRequestClientModel;
    /**
     * 
     * @type {ServiceRequestClientIdentificationModel}
     * @memberof ServiceRequestEditionModel
     */
    'clientIdentification'?: ServiceRequestClientIdentificationModel;
    /**
     * 
     * @type {Array<ContactInfoModel>}
     * @memberof ServiceRequestEditionModel
     */
    'clientContactInfo'?: Array<ContactInfoModel> | null;
    /**
     * 
     * @type {string}
     * @memberof ServiceRequestEditionModel
     */
    'formType'?: string | null;
    /**
     * 
     * @type {ServiceRequestCreationModelFormData}
     * @memberof ServiceRequestEditionModel
     */
    'formData'?: ServiceRequestCreationModelFormData;
    /**
     * 
     * @type {boolean}
     * @memberof ServiceRequestEditionModel
     */
    'ignoreBTypeVisibilityValidation'?: boolean;
    /**
     * 
     * @type {number}
     * @memberof ServiceRequestEditionModel
     */
    'id'?: number;
}